Add simple climber and simple slider
Added Simple Climber and Slider

Added Simple Slider

updated simple slider to use manual joystick input

Updated simple climber

Updated simple slider

+++ src/com/palyrobotics/frc2017/config/RobotState.java
@@ -21,7 +21,7 @@ public class RobotState {
 
 	// Pose stores encoder and gyro data for drivetrain
 	public Pose drivePose = new Pose(0,0,0,0,0,0);
-
+	
 	public double climberCurrentDraw;
-	public int climberEncoder;
+	public double climberEncoder;
 }
